% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_NOAA_ERA.R
\name{bind_NOAA_ERA}
\alias{bind_NOAA_ERA}
\title{Binds NOAA GEFS and ERA5 weather data.}
\usage{
bind_NOAA_ERA(start)
}
\arguments{
\item{start}{Start date of dates requested}
}
\value{
List of merged dataframes.
}
\description{
Merges NOAA GEFS and ERA5 site weather data by sites.
}
