% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5_Aggregate.R
\name{ERA5_aggregate}
\alias{ERA5_aggregate}
\title{Aggregate ERA5 to daily step}
\usage{
ERA5_aggregate(data)
}
\arguments{
\item{data}{List of site weather dataframes}

\item{ensemble_col}{column name with the ensembles}
}
\value{
Aggregated (6-hour) list of site weather dataframes
}
\description{
Aggregates NOAA GEFS weather variable data from a 6-hour to daily
time step.
}
