% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_request.R
\name{era5_request}
\alias{era5_request}
\title{Download ERA5 hourly weather data.}
\usage{
era5_request(site_coord, dates, fname = "download.nc")
}
\arguments{
\item{site_coord}{Site lon/lat vector (max lat, min lon, min lat, max lon)}

\item{dates}{Sequence of R "Date" class objects.}
}
\value{
A dataframe of the requested site variables.
}
\description{
Requests max temp, min temp, dew temp, solar rad, wind speed,
and precipitation data for the lon/lat sites provided.
}
