% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_tile_request.R
\name{era5_tile_read}
\alias{era5_tile_read}
\title{Reads ERA5 nc files that was already on the local disk}
\usage{
era5_tile_read(ncfile = "~/download.nc", sdate)
}
\arguments{
\item{ncfile}{ncfile}

\item{sdate}{Start date of hourly data}
}
\value{
A list of site var data frames
}
\description{
Reads ERA5 nc files that was already on the local disk
}
