% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_gefs_aggregate.R
\name{noaa_gefs_aggregate}
\alias{noaa_gefs_aggregate}
\title{Aggregate NOAA GEFS to 6-hour step}
\usage{
noaa_gefs_aggregate(data)
}
\arguments{
\item{data}{List of site weather dataframes}
}
\value{
Aggregated (6-hour) list of site weather dataframes
}
\description{
Aggregates NOAA GEFS weather variable data from a daily to 6-hour
time step.
}
