% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_gefs_request.R
\name{noaa_gefs_request}
\alias{noaa_gefs_request}
\title{Download NOAA GEFS data for a list of sites}
\usage{
noaa_gefs_request(sites)
}
\arguments{
\item{sites}{A dataframe of sites (ID, lat, lon)}
}
\value{
A list of site dataframes
}
\description{
Requests max temp, min temp, dew temp, solar rad, wind speed,
and precipitation data for the lon/lat sites provided.
}
