% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_met.R
\name{write_met}
\alias{write_met}
\title{Write the ERA5 met file to necdf for pSIMS}
\usage{
write_met(Metdata, filename = "test.nc4")
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Write the ERA5 met file to necdf for pSIMS
}
